pyproject.toml
[tool.poetry]
name = "simple_map"
version = "0.1.0"
description = ""
authors = ["Pedro Belo <pn999717@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
python-decouple = "^3.4"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
autopep8 = "^1.5.6"
django-extensions = "^3.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
